feat: Create visualizer controls
The VisualizerControls molecule component can be used to
control a visualizer. It consists of the following five
buttons:

 - shuffle
 - backward
 - play or pause (depends on the 'playing' prop)
 - forward
 - repeat

Each button is controlled by its corresponding action prop
(eg. onShuffle for the shuffle action) and it can disabled
by passing in the corresponding disabled prop.
(eg. shuffleDisabled for the shuffle button)