feat: add notCased prop to Button
The Button component previously displayed it's text in
uppercase. While this is a desirable behavior generally,
sometimes, the case should not be modified.

This may be achieved by passing in custom styles using the
className prop. However, the use case, I feel is common
enough to warrant an additional prop that will disable the
default transformation of the text to uppercase.