fix: Bars should remain in frame of the SortChart
The size of the margin of the Bar with the comparing class
is now calculated relative to the width of the bar. This is
visually more pleasing and solves the cases where the
combined width of the bars exceed the width of the
SortChart, thus causing some of the bars to be removed from
the visual frame of the SortChart.